requirements.txt
# Toolbox Everything — runtime dependencies
# Synchronisé avec les `import` réels du code (audit v1.3.1).
# Pour le dev : voir requirements-dev.txt

# --- Flask core ---
Flask==3.1.3
Werkzeug==3.1.8
python-dotenv==1.2.2
Flask-Compress==1.24

# --- Rate limiting (protection DoS sur endpoints serveur) ---
Flask-Limiter==4.1.1
redis==6.2.0

# --- Médias / images ---
Pillow==12.2.0
# FFmpeg est appelé directement via subprocess (cf. media_converter/routes.py),
# pas besoin du wrapper `ffmpeg-python`.

# --- Téléchargement vidéo / audio (multi-sources) ---
yt-dlp==2026.03.17

# --- HTTP client (healthcheck Stirling PDF) ---
requests==2.33.1

# --- Serveur WSGI prod + compression ---
gunicorn==25.3.0
brotli==1.2.0

# --- TLS / HTTP pinning (transitifs explicites pour la sécurité) ---
certifi>=2026.2.25
urllib3>=2.6.3
